+++ config/request-logger.php
@@ -78,4 +78,17 @@ return [
         'Authorization',
         'filter.search',
     ],
+
+    /*
+    |--------------------------------------------------------------------------
+    | User Model
+    |--------------------------------------------------------------------------
+    |
+    | This is the model used for user relationships.
+    | You can set a custom user model for relationships.
+    |
+    | Leaving this empty will use the model from the 'users' auth provider.
+    |
+    */
+    'user_model' => env('REQUEST_LOGGER_USER_MODEL'),
 ];